two way binding, child to parent component, lifting the state up
