+++ src/App.js
@@ -3,9 +3,14 @@ import ExpenseItems from './components/Expenses/ExpenseItems';
 import NewExpense from './components/Expenses/NewExpense';
 
 function App() {
+  const addExpenseHandler = expense => {
+    console.log('addExpenseHandler');
+    console.log(expense);
+  }
+
   return (
     <div>
-      <NewExpense />
+      <NewExpense addAddExpense={addExpenseHandler}/>
       <ExpenseItems />
     </div>
   );
