+++ src/css/global.scss
@@ -80,7 +80,7 @@ body {
   // NewExpenses.js
   .new-expense{
     background-color: #a892ee;
-    padding: 1rem;
+    padding: 2rem;
     margin: 2rem auto;
     width: 50rem;
     max-width: 95%;
@@ -116,14 +116,14 @@ body {
     }
   }
   .expense_form{
-    .new-expense__controls {
-      display: flex;
-      flex-wrap: wrap;
-      gap: 1rem;
+    .new-expenses_controls {
+      display: grid;
+      gap: 2rem;
       margin-bottom: 1rem;
       text-align: left;
+      grid-template-columns: repeat(3,minmax(0,1fr));
   
-      .new-expense__control{
+      .expenses_control{
         label {
           font-weight: bold;
           margin-bottom: 0.5rem;
